requirements.txt
asgiref==3.7.2
attrs==23.1.0
backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
dj-database-url==2.0.0
Django==4.0.4
django-cors-headers==4.1.0
django-environ==0.10.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.3
gunicorn==20.0.4
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
inflection==0.5.1
jsonschema==4.17.3
Markdown==3.4.3
Pillow==10.0.0
pkgutil_resolve_name==1.3.10
psycopg2==2.9.3
pyfcm==1.5.4
PyJWT==2.7.0
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.3
zipp==3.15.0
